fix: use int distribution in mmap test

+++ test/unit_tests/cpp/test_mmap.cpp
@@ -38,10 +38,10 @@ std::vector<uint8_t> MakeBinaryData(
         size_t seed) {
     std::vector<uint8_t> database(n * d);
     std::mt19937 rng(seed);
-    std::uniform_int_distribution<uint8_t> distrib(0, 255);
+    std::uniform_int_distribution<int> distrib(0, 255);
 
     for (size_t i = 0; i < n * d; i++) {
-        database[i] = distrib(rng);
+        database[i] = static_cast<uint8_t>(distrib(rng));
     }
     return database;
 }
